+++ src/function/arithmetic/mod.js
@@ -69,17 +69,11 @@ export const createMod = /* #__PURE__ */ factory(name, dependencies, ({ typed, c
       'number, number': _modNumber,
 
       'BigNumber, BigNumber': function (x, y) {
-        if (y.isNeg()) {
-          throw new Error('Cannot calculate mod for a negative divisor')
-        } return y.isZero() ? x : x.mod(y)
+        return y.isZero() ? x : x.sub(y.mul(floor(x.div(y))))
       },
 
       'Fraction, Fraction': function (x, y) {
-        if (y.compare(0) < 0) {
-          throw new Error('Cannot calculate mod for a negative divisor')
-        }
-        // Workaround suggested in Fraction.js library to calculate correct modulo for negative dividend
-        return x.compare(0) >= 0 ? x.mod(y) : x.mod(y).add(y).mod(y)
+        return y.equals(0) ? x : x.sub(y.mul(floor(x.div(y))))
       }
     },
     matrixAlgorithmSuite({